Crate homing

Structs

Bee ğŸ”’

bee struct to hold information about the snapshot and its position

Circle ğŸ”’

obstacle struct for circular objects

Grid ğŸ”’

Grid struct for all your grid needs origin of the Grid is always (0,0)

Image ğŸ”’

datastructure to hold the Segments this will be used for the snapshot and the image that is cast onto the retina

Segment ğŸ”’

datastructure for Segments on the image circle

Vec2 ğŸ”’

datastructure for 2d vectors

VectorField ğŸ”’

VectorField struct for storing all generated vectors

World ğŸ”’

World that holds obstacles and the grid the bee is allowed to be on

Traits

Distance ğŸ”’

Obstacle ğŸ”’

trait for obstacles all obstacles will have to implement this trait

Functions

main ğŸ”’